+++ CHANGELOG.md
@@ -12,12 +12,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Support custom body for binary media upload via POST
+- Set 'Accept' header for \$batch proxy request to "multipart/mixed"
+- Set missing response header 'content-transfer-encoding: binary'
 
 ## Version 1.4.52 - 2020-10-27
 
 ### Fixed
 
-- Log warning for change set order violation, instead returning an error response 
+- Log warning for change set order violation, instead returning an error response
 
 ## Version 1.4.51 - 2020-10-27